+++ libtitania-x3d/Titania/X3D/Browser/Geometry3D/QuadSphereOptions.h
@@ -138,7 +138,7 @@ private:
 	std::vector <int32_t>
 	createCoordIndex () const;
 
-	std::vector <Vector3f>
+	std::vector <Vector3d>
 	createPoints () const;
 
 	virtual
