+++ libtitania-x3d/Titania/X3D/Components/Rendering/IndexedLineSet.h
@@ -163,7 +163,7 @@ public:
 
 	///  @name Operations
 
-	std::vector <Vector3f>
+	std::vector <Vector3d>
 	getPolylines () const;
 
 	virtual
